LayoutTests/http/tests/security/drag-drop-local-file-expected.txt

CONSOLE MESSAGE: Not allowed to drag local resource: foobar
CONSOLE MESSAGE: PASS
Dragme
